change logic of applying offsets from NIDAQ analog output
since the offset to the soloist depends on the solenoid state (due to problems with the electronics) we apply a different offset to the output when the solenoid is on or off.

the AnalogOutput class write and set_to_idle methods receive arguments specifying the offset to apply

so to does the NI ao_write method

the Controller class now stores the two different offset values and sends the value depending on the state of the solenoid.

changing the state of the solenoid through the Controller also changes the value of the NIDAQ output